requirements.txt
streamlit
python-dotenv
streamlit
pandas
requests
openai
langchain
numpy
pylance
langchain-community